Added order details view for admin

+++ App/Controllers/Products.php
@@ -319,19 +319,21 @@ final class Products {
             DB_Model::roll_back();
             return Response::redirect('/');
         }
-        $res = DB_Model::query('delete from product_translations where product_id = :id')
-            ->bind_values(['id' => $id])
-            ->execute();
-        if (!$res->ok) {
-            DB_Model::roll_back();
-            return Response::redirect('/');
-        }
-        $res = DB_Model::query('delete from product_images where product_id = :id')
-            ->bind_values(['id' => $id])
-            ->execute();
-        if (!$res->ok) {
-            DB_Model::roll_back();
-            return Response::redirect('/');
+        if (DB_Model::$current_db === DB_Type::SQLITE) {
+            $res = DB_Model::query('delete from product_translations where product_id = :id')
+                ->bind_values(['id' => $id])
+                ->execute();
+            if (!$res->ok) {
+                DB_Model::roll_back();
+                return Response::redirect('/');
+            }
+            $res = DB_Model::query('delete from product_images where product_id = :id')
+                ->bind_values(['id' => $id])
+                ->execute();
+            if (!$res->ok) {
+                DB_Model::roll_back();
+                return Response::redirect('/');
+            }
         }
         DB_Model::commit();
